Alter the directory structure of the quotes plugin.
