+++ plugins/quotes/lib/quote_display.rb
@@ -1,0 +1,10 @@
+class QuoteDisplay
+  class << self
+
+    def output(m, quote)
+      output = quote.nil? || quote.empty? ? 'Quote added.' : quote
+      m.reply output
+    end
+
+  end
+end
